pyproject.toml
[project]
name = "node-drag-watcher"
version = "0.1.0"
description = "Watch OSM augmented diffs for accidental node drags"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.0",
    "requests>=2.31",
]

[project.scripts]
node-drag-watcher = "watcher:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
